VirtualMachineEngine::SingleClientSchedule
