+++ oneflow/core/vm/control_stream_type_test.cpp
@@ -45,7 +45,7 @@ TEST(ControlStreamType, new_symbol_symbol) {
   ASSERT_TRUE(vm->pending_msg_list().empty());
   CHECK_JUST(vm->Receive(&list));
   ASSERT_EQ(vm->pending_msg_list().size(), 1 * 2);
-  vm->Schedule();
+  vm->SingleClientSchedule();
   ASSERT_TRUE(vm->pending_msg_list().empty());
   ASSERT_TRUE(vm->active_stream_list().empty());
   ASSERT_EQ(vm->thread_ctx_list().size(), 1 * 2);
